Ticket #47378 - fix recent compiler warnings
https://fedorahosted.org/389/ticket/47378
Reviewed by: nhosoi (Thanks!)
Branch: master
Fix Description: Fix the various compiler warnings as reported by
http://jenkins.cloud.fedoraproject.org/job/389-ds-base/ws/build-warns.1370449480.txt
In addition, fixed a bug in the slapi_re_subs_ext - if there was an
ampersand that did not need to be replaced, that is, if the ampersand
was part of an LDAP search filter, it was replaced with a '('.  Also, an
ampersand preceeded by an open parenthesis and followed by a non-open
parenthesis is supposed to be replaced in the string, but this was not
working.
Platforms tested: RHEL6 x86_64
Flag Day: no
Doc impact: no
(cherry picked from commit 0dc230b007dc3ad292961f378421b03736b68f9e)

+++ lib/libaccess/acltools.cpp
@@ -1741,7 +1741,7 @@ static PList_t ACLAttr2IndexPList = NULL;
 int
 ACL_InitAttr2Index(void)
 {
-    int i;
+    intptr_t i;
 
     if (ACLAttr2IndexPList) return 0;
